MISRA Compliance Update (#121)
* Updated the MISRA.md and misra.config files after meeting with senior
SDE.
Put inline supression for a comparison related to the SIZE_MAX macro.
Want to get clarification about the line before putting a change in.

* Fixing some whitespace/formatting issues

* Changing MISRA.md file to reflect new format, modified inline supression in source file to match new formatting

* Adding words to lexicon, and fixing links

* Minor update to MISRA.md file to use an actual violation as the example, and expanding on a message

* Update source/core_json.c

Remove extra set of square brackets

Co-authored-by: Aniruddha Kanhere <60444055+AniruddhaKanhere@users.noreply.github.com>

* Changes to the way we determine the end in skipOneHexEscape()

* Removed a redundant check of a variable

* Formatting fix and adding a test in to get more line coverage

* Trying to reach 100% branch coverage

* Adding the removal of debug for coverity target, and then removing two rule exceptions from the misra.config due to the change

* skipOneHexEscape had a line that was flagged as a MISRA 14.3 rule
violation. It was flagged because it believed that the if statement
comparison was invariant. This could be proven as a bug by assigning the
variable a value larger than the comparison, and then still receiving
the violation. A logic change has been made to get around this, but it
now fails CBMC proofs. Trying a different if statement to see if this
passes checks.

* Forgot to add inital assign back in

* After a lot of attempts to create a MISRA and CBMC compliant version of
skipOneHexEscape() I believe proof was found that shows the MISRA
violation is a false flag. Due to this I believe that this should simply
receive a suppression and the focus should be on the CBMC proofs.

Co-authored-by: Soren Ptak <skptak@amazon.com>
Co-authored-by: Aniruddha Kanhere <60444055+AniruddhaKanhere@users.noreply.github.com>