Make coreJSON more verification friendly
In order to turn all bounded CBMC proofs into unbounded proofs,
we must make some small changes to coreJSON. Basically, we want
to avoid uninitialized variables (CBMC sees them as non-deterministic
values) and make sure the code in loops is resilient enough when
working with inputs of arbitrary size. The latter is important
because we must annotate all loops with loop invariants.
The implementation should work together with these annotations
in order to prove memory safety for inputs of arbitrary size.

Signed-off-by: Felipe R. Monteiro <felisous@amazon.com>